Unify MRZAnalyzer class
