--- core-lib/src/main/java/org/idpass/smartscanner/lib/mrz/analyzers/MRZAnalyzer.kt
@@ -1,116 +1,0 @@
-/*
- * Copyright (C) 2020 Newlogic Pte. Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- *
- *
- */
-package org.idpass.smartscanner.lib.mrz.analyzers
-
-import android.app.Activity
-import android.content.Intent
-import android.graphics.Bitmap
-import android.os.Bundle
-import android.util.Log
-import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.ImageProxy
-import com.google.gson.Gson
-import org.idpass.smartscanner.api.ScannerConstants
-import org.idpass.smartscanner.lib.SmartScannerActivity
-import org.idpass.smartscanner.lib.mrz.MRZCleaner
-import org.idpass.smartscanner.lib.mrz.MRZResult
-import org.idpass.smartscanner.lib.platform.extension.cacheImagePath
-import org.idpass.smartscanner.lib.platform.extension.cacheImageToLocal
-import org.idpass.smartscanner.lib.platform.extension.encodeBase64
-import org.idpass.smartscanner.lib.scanner.config.ImageResultType
-import org.idpass.smartscanner.lib.scanner.config.MrzFormat
-
-open class MRZAnalyzer(
-    private val activity: Activity,
-    private val intent : Intent,
-    private val imageResultType: String,
-    private val format: String?
-) : ImageAnalysis.Analyzer {
-
-    override fun analyze(imageProxy: ImageProxy) {}
-
-    fun processResult(mrz : String, bitmap : Bitmap, rotation : Int) {
-        val imagePathFile = activity.cacheImagePath()
-        bitmap.cacheImageToLocal(imagePathFile, rotation)
-        val imageString = if (imageResultType == ImageResultType.BASE_64.value) bitmap.encodeBase64(rotation) else imagePathFile
-        val result = when (format) {
-            MrzFormat.MRTD_TD1.value -> MRZResult.formatMrtdTd1Result(
-                MRZCleaner.parseAndCleanMrtdTd1(
-                    mrz
-                ), imageString
-            )
-            else -> MRZResult.formatMrzResult(
-                MRZCleaner.parseAndClean(mrz),
-                imageString
-            )
-        }
-        if (intent.action == ScannerConstants.IDPASS_SMARTSCANNER_MRZ_INTENT ||
-            intent.action == ScannerConstants.IDPASS_SMARTSCANNER_ODK_MRZ_INTENT) {
-            sendBundleResult(activity, intent, mrzResult = result)
-        }
-        else {
-            val jsonString = Gson().toJson(result)
-            sendAnalyzerResult(activity, jsonString)
-        }
-    }
-
-    private fun sendAnalyzerResult(activity: Activity, result: String? = null) {
-        val data = Intent()
-        Log.d(SmartScannerActivity.TAG, "Success from BARCODE")
-        Log.d(SmartScannerActivity.TAG, "value: $result")
-        data.putExtra(SmartScannerActivity.SCANNER_RESULT, result)
-        activity.setResult(Activity.RESULT_OK, data)
-        activity.finish()
-    }
-
-    private fun sendBundleResult(activity: Activity, intent : Intent, mrzResult : MRZResult? = null) {
-        val bundle = Bundle()
-        Log.d(SmartScannerActivity.TAG, "Success from MRZ")
-        mrzResult?.let { result ->
-            if (intent.action == ScannerConstants.IDPASS_SMARTSCANNER_ODK_MRZ_INTENT) {
-                bundle.putString(ScannerConstants.IDPASS_ODK_INTENT_DATA, result.documentNumber)
-            }
-            // TODO implement proper image passing
-            //  bundle.putString(ScannerConstants.MRZ_IMAGE, result.image)
-            bundle.putString(ScannerConstants.MRZ_CODE, result.code)
-            bundle.putShort(ScannerConstants.MRZ_CODE_1, result.code1 ?: -1)
-            bundle.putShort(ScannerConstants.MRZ_CODE_2, result.code2 ?: -1)
-            bundle.putString(ScannerConstants.MRZ_DATE_OF_BIRTH, result.dateOfBirth)
-            bundle.putString(ScannerConstants.MRZ_DOCUMENT_NUMBER, result.documentNumber)
-            bundle.putString(ScannerConstants.MRZ_EXPIRY_DATE, result.expirationDate)
-            bundle.putString(ScannerConstants.MRZ_FORMAT, result.format)
-            bundle.putString(ScannerConstants.MRZ_GIVEN_NAMES, result.givenNames)
-            bundle.putString(ScannerConstants.MRZ_SURNAME, result.surname)
-            bundle.putString(ScannerConstants.MRZ_ISSUING_COUNTRY, result.issuingCountry)
-            bundle.putString(ScannerConstants.MRZ_NATIONALITY, result.nationality)
-            bundle.putString(ScannerConstants.MRZ_SEX, result.sex)
-            bundle.putString(ScannerConstants.MRZ_RAW, result.mrz)
-        }
-        val result = Intent()
-        val prefix = if (intent.hasExtra(ScannerConstants.IDPASS_ODK_PREFIX_EXTRA)) {
-            intent.getStringExtra(ScannerConstants.IDPASS_ODK_PREFIX_EXTRA)
-        } else { "" }
-        result.putExtra(ScannerConstants.RESULT, bundle)
-        // Copy all the values in the intent result to be compatible with other implementations than commcare
-        for (key in bundle.keySet()) {
-            result.putExtra(prefix + key, bundle.getString(key))
-        }
-        activity.setResult(Activity.RESULT_OK, result)
-        activity.finish()
-    }
-}
